Fix #125
