+++ Assessments.Mapping/SpeciesAssessment2021Profile.cs
@@ -17,7 +17,7 @@ namespace Assessments.Mapping
                 .ForMember(dest => dest.PopulationScope, opt => opt.MapFrom(src => src.Omfang))
                 .ForMember(dest => dest.TimeScope, opt => opt.MapFrom(src => src.Tidspunkt))
                 .ForMember(dest => dest.Id, opt => opt.MapFrom(src => src.Id))
-                .AfterMap(CorrectImpactFactors);
+                .AfterMap(SpeciesAssessment2021ProfileHelper.CorrectImpactFactors);
 
             CreateMap<Rodliste2019.TrackInfo, SpeciesAssessment2021TrackInfo>();
             CreateMap<Rodliste2019.MinMaxProbable, SpeciesAssessment2021MinMaxProbable>();
@@ -156,212 +156,7 @@ namespace Assessments.Mapping
 
                 .ForMember(dest => dest.ReasonCategoryChange, opt => opt.MapFrom(src => src.ÅrsakTilEndringAvKategori))
                 .ForMember(dest => dest.ÅrsakTilNedgraderingAvKategori, opt => opt.MapFrom(src => src.ÅrsakTilNedgraderingAvKategori))
-                .AfterMap((src, dest) =>
-                {
-                    if (dest.Category == "LCº")
-                    {
-                        dest.CriteriaSummarized = string.Empty;
-                    }
-                });
+                .AfterMap((src, dest) => { SpeciesAssessment2021ProfileHelper.BlankCriteriaSumarizedBasedOnCategory(dest); });
         }
-
-        #region CorrectImpactFactors
-
-        private static Dictionary<string, string> _replaceTextDictionary = new()
-        {
-            {
-                "Skogbruk/avvirkning",
-                "Skogbruk (kommersielt)"
-            },
-            {
-                "Åpne hogstformer (flatehogst og frøtrehogst som også inkluderer uttak av rotvelt, råtne trær, tørrgran etc.)",
-                "Åpne hogstformer (flatehogst og frøtrestillingshogst som også inkluderer uttak av rotvelt, råtne trær, tørrgran etc.)"
-            },
-            {
-                "Habitatpåvirkning på ikke landbruksarealer (terrestrisk)",
-                "Habitatpåvirkning - ikke jord- eller skogbruksaktivitet (terrestrisk)"
-            },
-            {
-                "Oppdemming / vannstandsregulering / overføring av vassdrag",
-                "Oppdemming/vannstandsregulering/overføring av vassdrag"
-            },
-            {
-                "Tynning, vedhogst, avvirkning av spesielle type trær (gamle, hule, brannskade)",
-                "Vedhogst, avvirkning av spesielle type trær (gamle, hule, brannskade)"
-            }
-        };
-        private static Dictionary<string, string> _replaceIdDictionary = new()
-        {
-            {
-                "10.1",
-                "10."
-            },
-            {
-                "11.1",
-                "11."
-            },
-            {
-                "12.1",
-                "12."
-            },
-            {
-                "0.1",
-                "0."
-            },
-            {
-                "0.1.",
-                "0."
-            }
-        };
-
-        private static Dictionary<string, Tuple<string, string, string>> mapImpactfactor2010Dictionary =
-            new Dictionary<string, Tuple<string, string, string>>
-            {
-                {
-                    "1.1.2.Skogbruk/avvirkning",
-                    new Tuple<string, string, string>("1.1.2.1.",
-                        "Skogsdrift, hogst og skjøtsel",
-                        "Påvirkning på habitat > Landbruk > Skogbruk (kommersielt)")
-                },
-                {
-                    "1.1.2.1.Flatehogst",
-                    new Tuple<string, string, string>("1.1.2.1.1.",
-                        "Åpne hogstformer (flatehogst og frøtrehogst som også inkluderer uttak av rotvelt, råtne trær, tørrgran etc.)",
-                        "Påvirkning på habitat > Landbruk > Skogbruk (kommersielt) > Skogsdrift, hogst og skjøtsel")
-                },
-                {
-                    "1.1.2.2.Plukkhogst, tynning, vedhogst",
-                    new Tuple<string, string, string>("1.1.2.1.2",
-                        "Lukkede hogstformer (plukkhogst, skjermstilling, tynning, uttak av enkelttrær, inkludert uttak av rotvelt, råtne trær, tørrgran etc.)",
-                        "Påvirkning på habitat > Landbruk > Skogbruk (kommersielt) > Skogsdrift, hogst og skjøtsel")
-                },
-                {
-                    "1.1.2.3.Fjerning av dødt virke", new Tuple<string, string, string>("1.1.2.1.3.",
-                        "Ungskogrydding (rydding i ungskog)",
-                        "Påvirkning på habitat > Landbruk > Skogbruk (kommersielt) > Skogsdrift, hogst og skjøtsel")
-                },
-                //{
-                //    "1.2.4.1.Fjerning av dødt virke", new Tuple<string, string, string>("1.1.2.1.4.",
-                //    "Uttak av død ved (stående *gadd* og liggende *læger*)",
-                //    "Påvirkning på habitat > Habitatpåvirkning - ikke jord- eller skogbruksaktivitet (terrestrisk) > Annen påvirkning på habitat")
-                //},
-                {
-                    "4.2.Andre",
-                    new Tuple<string, string, string>("4.",
-                        "Tilfeldig mortalitet",
-                        "")
-                },
-            };
-
-        private static Dictionary<string, Tuple<string, string>> fixImpactfactor2010Dictionary =
-            new Dictionary<string, Tuple<string, string>>
-            {
-                {
-                    "1.1.2.1.",
-                    new Tuple<string, string>(
-                        "Skogsdrift, hogst og skjøtsel",
-                        "Påvirkning på habitat > Landbruk > Skogbruk (kommersielt)")
-                },
-                {
-                    "1.1.2.1.1.",
-                    new Tuple<string, string>(
-                        "Åpne hogstformer (flatehogst og frøtrehogst som også inkluderer uttak av rotvelt, råtne trær, tørrgran etc.)",
-                        "Påvirkning på habitat > Landbruk > Skogbruk (kommersielt) > Skogsdrift, hogst og skjøtsel")
-                },
-                {
-                    "1.1.2.1.2",
-                    new Tuple<string, string>(
-                        "Lukkede hogstformer (plukkhogst, skjermstilling, tynning, uttak av enkelttrær, inkludert uttak av rotvelt, råtne trær, tørrgran etc.)",
-                        "Påvirkning på habitat > Landbruk > Skogbruk (kommersielt) > Skogsdrift, hogst og skjøtsel")
-                },
-                {
-                    "1.1.2.1.3.", new Tuple<string, string>(
-                        "Ungskogrydding (rydding i ungskog)",
-                        "Påvirkning på habitat > Landbruk > Skogbruk (kommersielt) > Skogsdrift, hogst og skjøtsel")
-                },
-                {
-                    "1.2.4.1.", new Tuple<string, string>(
-                        "Uttak av død ved (stående *gadd* og liggende *læger*)",
-                        "Påvirkning på habitat > Habitatpåvirkning - ikke jord- eller skogbruksaktivitet (terrestrisk) > Annen påvirkning på habitat")
-                },
-                {
-                    "4.",
-                    new Tuple<string, string>(
-                        "Tilfeldig mortalitet",
-                        "")
-                }
-            };
-
-        private static string _uttakAvDødVedStåendeGaddOgLiggendeLæger = "Uttak av død ved (stående gadd og liggende læger)";
-        private static void CorrectImpactFactors(Rodliste2019.Pavirkningsfaktor src, SpeciesAssessment2021ImpactFactor dest)
-        {
-            var pOverordnetTittel = src.OverordnetTittel;
-            var pBeskrivelse = src.Beskrivelse;
-            var pId = src.Id;
-
-            // map old 2010 factors to correct id and text
-            foreach (var item in mapImpactfactor2010Dictionary.Where(item => item.Key == pId.Trim() + pBeskrivelse.Trim()))
-            {
-                //Console.WriteLine(assessment.VurdertVitenskapeligNavn);
-                pBeskrivelse = item.Value.Item2;
-                pId = item.Value.Item1;
-                pOverordnetTittel = item.Value.Item3;
-            }
-
-            // fix som old texts 
-            foreach (var item in fixImpactfactor2010Dictionary.Where(item => item.Key == pId.Trim()))
-            {
-                //Console.WriteLine(assessment.VurdertVitenskapeligNavn);
-                if (pBeskrivelse != item.Value.Item1)
-                {
-                    pBeskrivelse = item.Value.Item1;
-                }
-
-                //pId = item.Value.Item1;
-                if (pOverordnetTittel == item.Value.Item2)
-                {
-                    pOverordnetTittel = item.Value.Item2;
-                }
-            }
-
-            foreach (var item in _replaceTextDictionary.Where(item =>
-                pOverordnetTittel.IndexOf(item.Key, StringComparison.InvariantCulture) > -1))
-            {
-                pOverordnetTittel = pOverordnetTittel.Replace(item.Key, item.Value, StringComparison.InvariantCulture);
-            }
-
-            foreach (var item in _replaceTextDictionary.Where(item =>
-                pBeskrivelse.IndexOf(item.Key, StringComparison.InvariantCulture) > -1))
-            {
-                pBeskrivelse = pBeskrivelse.Replace(item.Key, item.Value, StringComparison.InvariantCulture);
-            }
-
-            foreach (var item in _replaceIdDictionary.Where(item => pId == item.Key))
-            {
-                pId = item.Value;
-            }
-
-            var level = dest.Id.Split(".", StringSplitOptions.RemoveEmptyEntries).Length;
-
-            if (pId == "1.1.2.1.4." || pId == "1.2.4.1.")
-            {
-                // fjern alternativ med * og " brukt i disse
-                pBeskrivelse = _uttakAvDødVedStåendeGaddOgLiggendeLæger;
-            }
-
-            // todo: kanskje bruke denne.... inneholder sti, men Factorpath er hele stien...
-            //var under = level > 1
-            //    ? string.Join(" > ", pOverordnetTittel.Split(" > ").Skip(1)) + " - " + pBeskrivelse
-            //    : pBeskrivelse;
-            
-            dest.Id = pId;
-            dest.Factor = pBeskrivelse;
-            dest.FactorPath =
-                pOverordnetTittel.Split(" > ").Where(x => !string.IsNullOrWhiteSpace(x))
-                    .ToArray(); //.Union(new[] { pBeskrivelse }).Where(x => !string.IsNullOrWhiteSpace(x)).ToArray();
-            dest.GroupingFactor = pOverordnetTittel == "" ? pBeskrivelse : pOverordnetTittel.Split(" > ")[0];
-        }
-
-        #endregion
     }
 }
